set custom validity to a generic error message
Instead of requiring the server to send pretty elaborate responses in
the case of a error so that we can set the custom validity correctly,
this PR will set it to a generic error message.

Since the application is expected to show their own message we can
assume that a generic error message is fine to show since they are ever
just shown in certain situations. The majority of instances the user
is shown a error it's the application crafted message.

This will simplify the component quite a bit and doesn't require very
certain JSON responses from the server.