set input as invalid on input
Before this change we would make the element invalid while a request was
being made. This was happening inside the debounced function. Which
means that in the 300ms from triggering a input event a user could
submit a form with the input as being valid even though it hadn't been
validated yet.

This change moves the functionality that makes the input invalid while
loading outside the debounced function so that this isn't possible
anymore.