Remove asynchronous validation message support
The setValidity function can be enhanced to accept a Promise in a future
branch. Because both the network request and error message handling are
async, they need to be coordinated with a state machine to remove the
possibility of races between them corrupting shared state like the
current abort controller.